fix: update Capstone submodule and build fixes

+++ 3rdparty/Capstone/src/arch/MOS65XX/MOS65XXDisassembler.c
@@ -203,7 +203,8 @@ static void fillDetails(MCInst *MI, struct OpInfo opinfo, int cpu_type)
 			break;
 		case MOS65XX_AM_IMM:
 			detail->mos65xx.operands[detail->mos65xx.op_count].type = MOS65XX_OP_IMM;
-			detail->mos65xx.operands[detail->mos65xx.op_count].imm = MI->Operands[0].ImmVal;
+			/* Decoded immediates contain one byte, or two in 65816 mode. */
+			detail->mos65xx.operands[detail->mos65xx.op_count].imm = (uint16_t)MI->Operands[0].ImmVal;
 			detail->mos65xx.op_count++;
 			break;
 		case MOS65XX_AM_ACC:
@@ -212,7 +213,7 @@ static void fillDetails(MCInst *MI, struct OpInfo opinfo, int cpu_type)
 			detail->mos65xx.op_count++;
 			break;
 		case MOS65XX_AM_REL: {
-			int value = MI->Operands[0].ImmVal;
+			int64_t value = MI->Operands[0].ImmVal;
 			if (MI->op1_size == 1)
 				value = 2 + (signed char)value;
 			else
@@ -226,7 +227,8 @@ static void fillDetails(MCInst *MI, struct OpInfo opinfo, int cpu_type)
 			int value =	3 + (signed char)MI->Operands[1].ImmVal;
 			/* BBR0, zp, rel  and BBS0, zp, rel */
 			detail->mos65xx.operands[detail->mos65xx.op_count].type = MOS65XX_OP_MEM;
-			detail->mos65xx.operands[detail->mos65xx.op_count].mem = MI->Operands[0].ImmVal;
+			/* The zero-page address is decoded from code[1]. */
+			detail->mos65xx.operands[detail->mos65xx.op_count].mem = (uint32_t)MI->Operands[0].ImmVal;
 			detail->mos65xx.operands[detail->mos65xx.op_count+1].type = MOS65XX_OP_MEM;
 			detail->mos65xx.operands[detail->mos65xx.op_count+1].mem = (MI->address + value) & 0xffff;
 			detail->mos65xx.op_count+=2;
@@ -235,7 +237,8 @@ static void fillDetails(MCInst *MI, struct OpInfo opinfo, int cpu_type)
 		default:
 			for (i = 0; i < MI->size; ++i) {
 				detail->mos65xx.operands[detail->mos65xx.op_count].type = MOS65XX_OP_MEM;
-				detail->mos65xx.operands[detail->mos65xx.op_count].mem = MI->Operands[i].ImmVal;
+				/* MOS65XX_getInstruction assembles at most a 24-bit address. */
+				detail->mos65xx.operands[detail->mos65xx.op_count].mem = (uint32_t)MI->Operands[i].ImmVal;
 				detail->mos65xx.op_count++;
 			}
 			break;
@@ -267,7 +270,8 @@ void MOS65XX_printInst(MCInst *MI, struct SStream *O, void *PrinterInfo)
 			break;
 	}
 
-	value = MI->Operands[0].ImmVal;
+	/* Decoded operands are unsigned values of at most three bytes. */
+	value = (unsigned int)MI->Operands[0].ImmVal;
 
 	switch (opinfo.am) {
 		default:
